Group Summaries/requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
httpx>=0.27.0
selenium>=4.15.0
webdriver-manager>=4.0.0
